+++ packages/api/src/controllers.rs
@@ -4,11 +4,13 @@ pub mod method_not_allowed;
 pub mod not_found;
 pub mod sample;
 pub mod totp;
+pub mod write;
 
 pub use delete_many::delete_many;
 pub use health_check::health_check;
 pub use method_not_allowed::method_not_allowed;
 pub use not_found::not_found;
 pub use sample::sample;
+pub use write::write;
 
 pub use totp::create_totp;
